strands-py/backend/requirements.txt
strands-agents>=1.30.0
strands-agents-tools>=0.2.22
boto3>=1.42.71
python-dotenv>=1.2.1
fastapi>=0.115.0
python-multipart>=0.0.2
docling>=2.70.0
pyyaml>=6.0.2
pytest>=9.0.0
httpx>=0.27.0